+++ site/source/components/Simulation/ObjectifDeSimulation.stories.tsx
@@ -2,7 +2,7 @@ import { Meta, StoryObj } from '@storybook/react'
 import { Option } from 'effect'
 
 import { BlocMontants } from '@/components/Simulateur/BlocMontants'
-import { eurosParAn } from '@/domaine/Montant'
+import { eurosParAn } from '@/domaine/MontantRécurrent'
 
 import { ObjectifDeSimulation } from './ObjectifDeSimulation'
