+++ site/source/contextes/comparateur/domaine/ModeleAssimileSalarie.test.ts
@@ -1,8 +1,8 @@
 import * as O from 'effect/Option'
 import { describe, expect, it } from 'vitest'
 
-import { eurosParAn, eurosParMois, toEurosParAn } from '@/domaine/Montant'
-import { plus } from '@/domaine/MontantRécurrent'
+import { toEurosParAn } from '@/domaine/Montant'
+import { eurosParAn, eurosParMois, plus } from '@/domaine/MontantRécurrent'
 
 import { ModèleAssimiléSalarié } from './ModeleAssimileSalarie'
