+++ site/source/contextes/comparateur/domaine/modeleComparable.ts
@@ -4,7 +4,8 @@ import Engine from 'publicodes'
 import { ReactNode } from 'react'
 
 import { StatutType } from '@/components/StatutTag'
-import { Montant, MontantRécurrent } from '@/domaine/Montant'
+import { Montant } from '@/domaine/Montant'
+import { MontantRécurrent } from '@/domaine/MontantRécurrent'
 import { DottedName } from '@/domaine/publicodes/DottedName'
 import { NomModèle } from '@/domaine/PublicodesSimulationConfig'
 import { Quantité } from '@/domaine/Quantite'
