+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/estActiviteProfessionnelle.ts
@@ -1,8 +1,7 @@
 import { Either, pipe } from 'effect'
 
 import { PLAFOND_ANNUEL_SECURITE_SOCIALE } from '@/domaine/ConstantesSociales'
-import { eurosParAn } from '@/domaine/Montant'
-import { estPlusGrandOuÉgalÀ } from '@/domaine/MontantRécurrent'
+import { estPlusGrandOuÉgalÀ, eurosParAn } from '@/domaine/MontantRécurrent'
 import { valeurCourante } from '@/domaine/ValeurAnnuelle'
 
 import { AffiliationObligatoire } from './erreurs'
