+++ site/source/contextes/économie-collaborative/domaine/location-de-meublé/régime-micro-entreprise.ts
@@ -1,10 +1,11 @@
 import { Either, Option, pipe } from 'effect'
 
 import { evalueAvecPublicodes } from '@/domaine/engine/engineSingleton'
-import { eurosParAn, Montant } from '@/domaine/Montant'
+import { Montant } from '@/domaine/Montant'
 import {
 	estPlusGrandOuÉgalÀ,
 	estPlusGrandQue,
+	eurosParAn,
 } from '@/domaine/MontantRécurrent'
 import {
 	AutoEntrepreneurChiffreAffaireDansPublicodes,
