+++ site/source/domaine/MontantPonctuel.ts
@@ -5,6 +5,12 @@ import { UnitéMonétairePonctuelle } from './Unites'
 
 export type MontantPonctuel = Montant<UnitéMonétairePonctuelle>
 
+export const euros = (valeur: number): Montant<'€'> => montant(valeur, '€')
+
+export const eurosParTitreRestaurant = (
+	valeur: number
+): Montant<'€/titre-restaurant'> => montant(valeur, '€/titre-restaurant')
+
 export const moins = dual<
 	<U extends UnitéMonétairePonctuelle>(
 		montantB: Montant<U>
