+++ site/source/pages/assistants/choix-du-statut/rémunération.tsx
@@ -6,7 +6,8 @@ import { useDispatch } from 'react-redux'
 import { usePersistingState } from '@/components/utils/persistState'
 import { Body, H3, InfoButton, MontantField, Strong } from '@/design-system'
 import { ValeurPublicodes } from '@/domaine/engine/PublicodesAdapter'
-import { eurosParAn, Montant } from '@/domaine/Montant'
+import { Montant } from '@/domaine/Montant'
+import { eurosParAn } from '@/domaine/MontantRécurrent'
 import { DottedName } from '@/domaine/publicodes/DottedName'
 import { enregistreLesRéponsesAuxQuestions } from '@/store/actions/actions'
 import { debounce } from '@/utils'
