+++ site/source/pages/simulateurs/comparaison-statuts/Objectifs.tsx
@@ -7,7 +7,7 @@ import {
 } from '@/components/Simulation/ObjectifSaisissableDeSimulation'
 import { useComparateur } from '@/contextes/comparateur'
 import { MontantField } from '@/design-system'
-import { MontantRécurrent } from '@/domaine/Montant'
+import { MontantRécurrent } from '@/domaine/MontantRécurrent'
 
 import { ChoixImposition } from './components/ChoixImposition'
 import { ChoixVersementLibératoire } from './components/ChoixVersementLiberatoire'
